Fix offer user

+++ app/src/Controller/OfferController.php
@@ -2,22 +2,16 @@
 
 namespace App\Controller;
 
-
-use App\Entity\Tag;
-use App\Entity\User;
-use App\Entity\Brand;
 use App\Entity\Offer;
 
 use App\Form\OfferType;
-use App\Entity\Application;
 use App\Form\ApplicationType;
-use App\Repository\OfferRepository;
 use App\Repository\BrandRepository;
+use App\Repository\OfferRepository;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 
 
 /**
@@ -34,43 +28,48 @@ class OfferController extends AbstractController
      * @Route("/", name="index", methods={"GET"})
      */
 
-    public function index(): Response
+    public function index(BrandRepository $brandRepository)
     {
         $repository = $this->getDoctrine()->getRepository(Offer::class);
 
+        $user = $this->getUser();
+        $brand = $brandRepository->findOneBy(['UserId' => $user]);
+
         $offer = $repository->findBy([
             'status' => 'Libre',
-        ]);
+        ], ['dateCreation' => 'DESC']);
+
         return $this->render('offer/index.html.twig', [
             'offers' =>  $offer,
+            'brand' => $brand
         ]);
     }
 
-
     /**
      * @Route("/new", name="new", methods={"GET", "POST"})
+     * @IsGranted("ROLE_MARQUE", statusCode=404, message="Vous n'avez pas accès à cette page!")
      */
 
-    public function new(Request $request)
+    public function new(Request $request, BrandRepository $brandRepository)
     {
 
         $offer = new Offer();
 
-        $dateStart = $offer->getDateStart();
-        $dateEnd = $offer->getDateEnd();
-
         $user = $this->getUser();
+
+        $brandId = $brandRepository->findOneBy(['UserId' => $user]);
 
         if (array_search("ROLE_MARQUE", $user->getRoles()) !== false) {
         }
-        $dateStart = $offer->getDateStart();
-        $dateEnd = $offer->getDateEnd();
 
         $form = $this->createForm(OfferType::class, $offer);
 
         $form->handleRequest($request);
 
+        $user = $this->getUser();
+
         if ($form->isSubmitted() && $form->isValid()) {
+            $offer->setBrandId($brandId);
             $em = $this->getDoctrine()->getManager();
             $em->persist($offer);
             $em->flush();
@@ -82,50 +81,59 @@ class OfferController extends AbstractController
 
 
         return $this->render('offer/new.html.twig', [
-
             'form' => $form->createView()
-
         ]);
     }
 
     /**
      * @Route("/show/{id}", name="show", methods={"GET"})
      */
-    public function show(Offer $offer): Response
+    public function show($id, Offer $offer, BrandRepository $brandRepository, OfferRepository $offerRepository)
     {
+        $offer = $offerRepository->find($id);
+
+        $user = $this->getUser();
+
+        $brand = $brandRepository->findOneBy(['UserId' => $user]);
+
         return $this->render('offer/show.html.twig', [
-            'offer' => $offer
+            'offer' => $offer,
+            'brand' => $brand
         ]);
     }
 
 
     /**
      * @Route("/edit/{id}", name="edit", methods={"GET", "POST"})
+     * @IsGranted("ROLE_MARQUE", statusCode=404, message="Vous n'avez pas accès à cette page!")
      */
-    public function edit(Offer $offer, Request $request)
+    public function edit($id, Offer $offer, Request $request, BrandRepository $brandRepository, OfferRepository $offerRepository)
     {
-        $dateStart = $offer->getDateStart();
-        $dateEnd = $offer->getDateEnd();
+        $offerId = $offerRepository->find($id);
+
+        $this->denyAccessUnlessGranted('CAN_EDIT', $offerId, "Vous n'avez pas acces");
+
         $form = $this->createForm(OfferType::class, $offer);
 
         $form->handleRequest($request);
+
         if ($form->isSubmitted() && $form->isValid()) {
             $em = $this->getDoctrine()->getManager();
             $em->flush();
 
             $this->addFlash('blue', 'Modification réussie');
 
-            return $this->redirectToRoute('offer_edit', ['id' => $offer->getId()]);
+            return $this->redirectToRoute('offer_index');
         }
 
         return $this->render('offer/edit.html.twig', [
             'offer' => $offer,
-            'form' => $form->createView()
+            'form' => $form->createView(),
         ]);
     }
 
     /**
-     * @Route("/apply/{id}/", name="apply", methods={ "GET", "POST"})     * 
+     * @Route("/apply/{id}/", name="apply", methods={ "GET", "POST"})
      */
     public function apply(Offer $offer, Request $request)
     {
@@ -154,15 +162,22 @@ class OfferController extends AbstractController
 
     /**
      * @Route("/delete/{id}/{token}", name="delete", methods={"GET"})
+     * @IsGranted("ROLE_MARQUE", statusCode=404, message="Vous n'avez pas accès à cette page!")
      */
-    public function delete(Offer $offer, $token)
+    public function delete($id, Offer $offer, $token, OfferRepository $offerRepository)
     {
+        $offerId = $offerRepository->find($id);
+        // dd($offer);
+        $this->denyAccessUnlessGranted('CAN_DELETE', $offerId, "Vous n'avez pas accès");
+
+
         if (!$this->isCsrfTokenValid('delete_offer' . $offer->getId(), $token)) {
             throw new \Exception('Token CSRF invalid');
         }
 
         $em = $this->getDoctrine()->getManager();
         $em->remove($offer);
+        $offer->setBrandId(null);
         $em->flush();
 
         $this->addFlash('red', 'Suppression réussie');
